Add implementation of a doubly circular linked list in C++
This commit adds a C++ implementation of a doubly circular linked list.
Theimplementation includes a Node class representating a node of the
list and a DoubluLinkedList class representating the list itself.
Important functions such as insertion at the beginning and end, clearing
the list, and displaying the elements are included. The implementation
is tested in the main function with various operations.